Refactor name RPC tests for test_framework change.
Update the name RPC tests for the recent test_framework change that was
made upstream.  Move also names.py to the test_framework
directory/module.
