+++ qa/rpc-tests/name_list.py
@@ -5,14 +5,8 @@
 
 # RPC test for name_list.
 
-# Add python-bitcoinrpc to module search path:
-import os
-import sys
-sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "python-bitcoinrpc"))
-
-from bitcoinrpc.authproxy import JSONRPCException
-from names import NameTestFramework
-from util import assert_equal
+from test_framework.names import NameTestFramework
+from test_framework.util import *
 
 class NameListTest (NameTestFramework):
