+++ qa/rpc-tests/name_multisig.py
@@ -5,15 +5,10 @@
 
 # RPC test for multisig handling with names.
 
-# Add python-bitcoinrpc to module search path:
-import os
-import sys
-sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "python-bitcoinrpc"))
+from test_framework.names import NameTestFramework
+from test_framework.util import *
 
-from bitcoinrpc.authproxy import JSONRPCException
 from decimal import Decimal
-from names import NameTestFramework
-from util import assert_equal
 import binascii
 
 class NameMultisigTest (NameTestFramework):
