+++ qa/rpc-tests/name_reorg.py
@@ -6,14 +6,8 @@
 # Test that reorgs (undoing) work for names.  This also checks that
 # cleaning the mempool with respect to conflicting name registrations works.
 
-# Add python-bitcoinrpc to module search path:
-import os
-import sys
-sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "python-bitcoinrpc"))
-
-from bitcoinrpc.authproxy import JSONRPCException
-from names import NameTestFramework
-from util import assert_equal
+from test_framework.names import NameTestFramework
+from test_framework.util import *
 
 class NameRegistrationTest (NameTestFramework):
