+++ qa/rpc-tests/name_scanning.py
@@ -5,14 +5,8 @@
 
 # RPC test for the name scanning functions (name_scan and name_filter).
 
-# Add python-bitcoinrpc to module search path:
-import os
-import sys
-sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "python-bitcoinrpc"))
-
-from bitcoinrpc.authproxy import JSONRPCException
-from names import NameTestFramework
-from util import assert_equal
+from test_framework.names import NameTestFramework
+from test_framework.util import *
 
 class NameScanningTest (NameTestFramework):
